Add configuration for use_git_ls_files
This allows us to decide which algorithm to use for finding code
ownership files within the repo.

If the configuration directive is set to use_git_ls_files then we will
use that rather than finding files via a glob. This will lead to a
significant speed improvement as finding files via glob is slow.