Fix JSON spec
We were checking for specific text in the JSONParseError, but when json was updated, the message changed. We really only care that there was a JSON parse error so we're shortening the string we're checking. This change will make it so we don't run into this error again in future version updates.